Enemies now are damaged and die from an EnemyController script
The EnemyController script is going to replace functions inside the specialised enemy scripts which are used in all enemies. This will reduce code and make things easier to edit. It will also make computing more efficient as it will get reference to the player once rather than on every spawn.